app/requirements.txt
nltk
cassandra-driver
venv-pack
pyarrow
fastparquet
pathvalidate
pyspark
tqdm
cqlsh